requirements.txt
streamlit==1.37.1
spacy==3.7.4
pymupdf==1.24.8
python-docx==1.1.2
pandas==2.2.2
numpy==1.26.4
unidecode==1.3.8
